Made ParallelLabelAction private and moved to ParallelStepExecution
